mlops_project1/pyproject.toml
[tool.poetry]
name = "mlops-labs2024"
version = "0.0.1"
description = "A short description of the project."
authors = ["fiesta1212"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
awscli = "^1.36.3"
black = "^24.10.0"
flake8 = "^7.1.1"
ipython = "^8.29.0"
isort = "^5.13.2"
jupyterlab = "^4.3.0"
loguru = "^0.7.2"
matplotlib = "^3.9.2"
mkdocs = "^1.6.1"
notebook = "^7.2.2"
numpy = "^2.1.3"
pandas = "^2.2.3"
pip = "^24.3.1"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.2"
tqdm = "^4.67.0"
typer = "^0.13.0"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
boto3 = "^1.35.62"
types-boto3 = "^1.0.2"
mlflow = "^2.18.0"
omegaconf = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
